% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic.R
\name{acoustic}
\alias{acoustic}
\alias{le2ts}
\alias{sA2sa}
\alias{sV2sv}
\alias{sigma2ts}
\alias{ts2le}
\alias{ts2sigma}
\alias{z_le2ts}
\title{Acoustic calculations}
\usage{
le2ts(le, species, dead = 0)

ts2sigma(ts, type = "sp")

ts2le(ts, species, dead = 0)

sigma2ts(sigma)

z_le2ts(le, z, species, dead = 0)

sA2sa(sA)

sV2sv(sV)
}
\arguments{
\item{le}{length (cm)}

\item{species}{MRI species code (31:capelin, 30:herring, etc)}

\item{dead}{small value added to \code{le} related to measurement method}

\item{ts}{target strength (dB re 1 m^2)}

\item{type}{type of cross-section, \code{sp} spherical, \code{bs} backscattering}

\item{sigma}{cross-section (m^2)}

\item{z}{depth (m)}

\item{sA}{}

\item{Sv}{}
}
\value{
\itemize{
  \item \code{le2ts}:
  \item \code{ts2sigma}:
  \item \code{ts2le}:
  \item \code{sigma2ts}:
  \item \code{sA2sa}:
  \item \code{sV2sv}:
}
}
\description{
\itemize{
  \item \code{le2ts}: Convert length to target strength
  \item \code{ts2sigma}: Convert target strength to cross-section
  \item \code{ts2le}: Convert target strength to length
  \item \code{sigma2ts}: convert cross-section to target strength.
  \item \code{sA2sa}: Convert NASC (sA) to area backscattering coefficient (sa or ABC)
  \item \code{sV2sv}: Convert mean volum backscattering strength (MVBS) to volume backscattering coefficent
}
}
\examples{
# fish length distribution, rounded to the next cm below:
lengths <- floor(rnorm(100, 15, 2))
# cross-section of capelin
sigmas <- ts2sigma(le2ts(lengths, 31, dead = 0.5))
# nautical area scattering coefficient ('back-scattering area density'):
NASC <- 10000
# area density of fish in millions per nmi^2
NASC/mean(sigmas)/1e6

}
\references{
MacLennan, D. N., Fernandes, P. G., and Dalen, J. 2002. 
A consistent approach to definitions and symbols in fisheries acoustics. 
ICES Journal of Marine Science, 59: 365–369.
}

