% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_stats}
\alias{boot_stats}
\title{Compute summary stats of a bootstrap vector}
\usage{
boot_stats(x)
}
\arguments{
\item{x}{A vector, e.g. of bootstrap replicates}
}
\value{
A data frame with mean, CV and selected percentiles.
}
\description{
Compute summary stats of a bootstrap vector
}
\examples{

  x <- sample(1:10)
  boot_stats(x)

}
\keyword{arith}
