% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{dist_mat}
\alias{dist_mat}
\title{Create distance matrix between two sets of positions with 'geo::arcdist'.}
\usage{
dist_mat(lat, lon, lat1 = NULL, lon1 = NULL)
}
\arguments{
\item{lat, }{lon, lat1, lon1 Positions according to the 'geo' paradigm.}
}
\description{
Create distance matrix between two sets of positions with 'geo::arcdist'.
}
\keyword{arith}

