% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionechi.R
\docType{data}
\name{fishData}
\alias{fishData}
\title{Biological variables of 3300 capelin.}
\format{A data frame with 3297 rows and 14 variables}
\source{
MRI, to be made available at \url{http://data.hafro.is/acoustics}
}
\usage{
data(fishData)
}
\description{
A dataset containing the biological variables and other attributes 
of almost 3300 capelin along with the part of total numbers and 
biomass estimates each sampled fish represents.
The variables are as follows:
}
\details{
\itemize{
  \item \code{stod} station number
  \item \code{sid} sample id
  \item \code{teg} species code, 31 is capelin
  \item \code{nr} number of fish in sample of upto 100 fish
  \item \code{l} length (cm) to the next cm below
  \item \code{w} weight (g), rounded to the closest 0.1 g
  \item \code{s} sex, 1 male, 2 female
  \item \code{a} age in years
  \item \code{m} maturity, 0 immature, 1 mature
  \item \code{TS} target strength (dB re 1 m^2)
  \item \code{sigma} acoustic cross section (m^2)
  \item \code{area} area id
  \item \code{nFish} number of fish in total the sampled fish represents
  \item \code{bFish} biomass (in g) of fish in total the sampled fish represents
}
}
\keyword{datasets}

