% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alks.R
\name{fix_missing_age}
\alias{fix_missing_age}
\title{Fix missing age.}
\usage{
fix_missing_age(fishData)
}
\arguments{
\item{fishData}{A data frame of observations of fish in otolith sample
with columns 'l' for length and 'a' for age.}
}
\description{
Given a set of observations of fish length and age, some ages missing,
predict the missing once from observed age group proportions by length.
}
\examples{
\dontrun{
  fishData <- fix_missing_age(fishData)
  }
}
\keyword{age,}
\keyword{length}
