% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geoneighbourID}
\alias{geoneighbourID}
\title{For two sets of postions, find index of closest in second set to each in first with 'geo::dist_mat'.}
\usage{
geoneighbourID(lat, lon, lat1 = NULL, lon1 = NULL)
}
\arguments{
\item{lat, }{lon, lat1, lon1 Positions according to the 'geo' paradigm.}
}
\description{
For two sets of postions, find index of closest in second set to each in first with 'geo::dist_mat'.
}
\keyword{arith}
