% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alks.R
\name{make_alk}
\alias{make_alk}
\title{Make age-length key (alk)}
\usage{
make_alk(alk_data)
}
\arguments{
\item{alk_data}{Data frame with biological data from a set
of trawl samples to be used as basis for the age-length key.}
}
\value{
An alk-tibble
}
\description{
Given fish data with age and length, tablulate and prepare an age-length key
}
\details{
First implementation bin-ns the lengths on the centimeter below.
}
\examples{
\dontrun{fishData \%>\%
  filter(!is.na(a)) -> alk_data
alk <- make_alk(alk_data)
  }
}
\keyword{alk}
