% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alks.R
\name{make_my_age}
\alias{make_my_age}
\title{Split a length distribution according to an age-length-key.}
\usage{
make_my_age(Nl, alk)
}
\arguments{
\item{Nl}{Named vector giving the length distribution.}

\item{alk}{Named matrix age length key (length by age in spite of the name).}
}
\description{
Split a length distribution according to an age-length-key. Uses utilities
\code{\link{lub}} and \code{\link{glb}} to establish an age distribution
for lengths missing from the alk.
}

