% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{nearest_neighbour_distance}
\alias{nearest_neighbour_distance}
\title{For two sets of postions, find closest distance to a pos in second set for each in first with 'geo::dist_mat'.}
\usage{
nearest_neighbour_distance(lat, lon, lat1 = NULL, lon1 = NULL)
}
\arguments{
\item{lat, }{lon, lat1, lon1 Positions according to the 'geo' paradigm.}
}
\description{
For two sets of postions, find closest distance to a pos in second set for each in first with 'geo::dist_mat'.
}
\keyword{arith}
