% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_stock_para.R
\name{pick_stock_para}
\alias{pick_stock_para}
\title{Aggregate numbers and biomass of selected stock components}
\usage{
pick_stock_para(df)
}
\arguments{
\item{df}{A data frame/tibble with numbers and biomass by:

  \itemize{
    \item Area, \code{area}, 
    \item Length, \code{l}, 
    \item Age, \code{a} and
    \item Maturity, \code{m}. 
  }
grouped by area.}
}
\value{
A data frame/tibble of the selected parameters as columns by area.
}
\description{
Aggregate numbers and biomass of selected stock components
}
\keyword{arith}
