% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{pool_estimates}
\alias{pool_estimates}
\title{A rough gauge of the combination of repeat coverages}
\usage{
pool_estimates(estimates, CVs)
}
\arguments{
\item{estimates, CVs}{Vectors of means and CVs in repeat coverages.}
}
\value{
Vector of estimated mean and CV of coverages combined.
}
\description{
A rough gauge of the combination of repeat coverages
}
\examples{

  ests <- c(570, 720)
  CVs <- c(0.2, 0.2)
  pool_estimates(ests, CVs)

}
\keyword{arith}

