% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alks.R
\name{predict_age}
\alias{predict_age}
\title{Predict age}
\usage{
predict_age(l, alk)
}
\arguments{
\item{l}{Fish length}

\item{alk}{An alk-tibble with cum age group props by length}
}
\value{
Stochastic age
}
\description{
Given fish length, predict age from empirical age group propotions.
}
\details{
For a single fish at a time, used inside a 'purrr::map_dbl'-call
}
\examples{
\dontrun{fishData \%>\%
  filter(!is.na(a)) -> alk_data
alk <- make_alk(alk_data)
set.sed(2345)
predict_age(15,alk)
predict_age(15,alk)
set.seed(2345)
predict_age(15,alk)
purrr::map_dbl(9:19,predict_age,alk=alk)
  }
}
\keyword{alk}
