% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_compact_table}
\alias{read_compact_table}
\title{Read BI500 Report.CompactTable}
\usage{
read_compact_table(report = "report.CompactTables", species = "CAPEL")
}
\arguments{
\item{report}{Report name, defaults to 'Report.CompactTables'.}
}
\description{
Read BI500 Report.CompactTable
}
\examples{

  
\dontrun{
  filename <- "/u2/reikn/Tac/2011/31-capelin/winterCapelin/gildi/2011/A01/Report.CompactTables"
  a01 <- read_compact_table(filename)
}
}
\keyword{manip}
