% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpx_rte.R
\name{read_gpx_rte}
\alias{read_gpx_rte}
\title{Read routepoints from gpx-file}
\usage{
read_gpx_rte(gpx_fn)
}
\arguments{
\item{gp_fn}{Path and file name of gpx-file}
}
\description{
Read routepoints from gpx-file, with sf::st_read and sf_st_coordiantes. 
In the context of capelin acoustic estimation used for reading in a perimeter 
of the capelin distribution area..
}
