% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionechi.R
\docType{data}
\name{rep}
\alias{rep}
\title{NASC in 0.1 nmi bins/EDSUs.}
\format{
A data frame with 34352 rows and 11 variables
}
\source{
MRI, to be made available at \url{http://data.hafro.is/acoustics}
}
\usage{
data(rep)
}
\description{
A dataset containing the survey track and acoustic density attributed to capelin
in 0.1 nautical mile (nmi) bins.
}
\details{
\itemize{
  \item \code{ship,nat,surv} Ship, nation and survey as IDed in LSSS
  \item \code{date,time} Start date and time of EDSU in some format
  \item \code{logstart,logstop} Log distance at beginning and end of EDSU/bin
  \item \code{lon,lat} postion in decimal degrees longitude and latitude
  \item \code{depth} depth (md
  \item \code{sa} NASC (m^2/nmi^2)
}
}
\keyword{datasets}
