% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_byla.R
\name{tabulate_byla}
\alias{tabulate_byla}
\title{Tabulate abundance by length and age with additional info on the margin}
\usage{
tabulate_byla(fishData, nScale = 1e+06, bScale = nScale)
}
\arguments{
\item{fishData}{data.frame of numbers and biomass by length and age}

\item{nScale, }{bScale scaling factor, default one million}
}
\value{
List of matrices of abundance by length and age hybridized with other info
on the margin, for the stock combined (\code{all}) and immature (\code{imm}) 
and mature (\code{mat}) parts of the stock.
}
\description{
Tabulate abundance by length and age with additional marginal info consisting of:
\itemize{
  \item Marginal sums of abundance and biomass for length categories and age groups
  \item Mean length at age and mean weight at length and age
  \item Percentage age distribution
}
}
\details{
Table traditionally included in the reports of the Northern Pelagic 
and Blue Whiting and the North Western Working Groups of ICES 
as well as in survey reports.
}
\examples{
data(fishData)
tabulate_byla(fishData)

}

