% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2gpx_trk.R
\name{tbl2gpx_trk}
\alias{tbl2gpx_trk}
\title{Write table coordinates out as track in a gpx-file}
\usage{
tbl2gpx_trk(tbl, fn)
}
\arguments{
\item{tbl}{Table with postions given as 'lat' and 'lon'}

\item{fn}{Filename to write gpx to.}
}
\description{
Write lat and lon in a table out as track in a gpx-file
using sf-utilities.
}
