% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bionechi.R
\docType{package}
\name{bionechi}
\alias{bionechi}
\alias{bionechi-package}
\title{bionechi}
\description{
A collection of functions having to do with acoustics and
estimation of numbers and biomass and presentation of results
from acoustic surveys as performed at the MRI, Reykjvík, Iceland.
}

